Search for nvidia-container-cli in the custom root directory
